Move utility methods to be string extensions
